// constants/gics/types.ts
export type GicsSector =
  | "Communication Services"
  | "Consumer Discretionary"
  | "Consumer Staples"
  | "Energy"
  | "Financials"
  | "Health Care"
  | "Industrials"
  | "Information Technology"
  | "Materials"
  | "Real Estate"
  | "Utilities";

export type GicsIndustryGroup =
  | "Automobiles & Components"
  | "Banks"
  | "Capital Goods"
  | "Commercial & Professional Services"
  | "Consumer Discretionary Distribution & Retail"
  | "Consumer Durables & Apparel"
  | "Consumer Services"
  | "Consumer Staples Distribution & Retail"
  | "Energy"
  | "Equity Real Estate Investment Trusts (REITs)"
  | "Financial Services"
  | "Food, Beverage & Tobacco"
  | "Health Care Equipment & Services"
  | "Household & Personal Products"
  | "Insurance"
  | "Materials"
  | "Media & Entertainment"
  | "Pharmaceuticals, Biotechnology & Life Sciences"
  | "Real Estate Management & Development"
  | "Semiconductors & Semiconductor Equipment"
  | "Software & Services"
  | "Technology Hardware & Equipment"
  | "Telecommunication Services"
  | "Transportation"
  | "Utilities";
// ...
export type GicsIndustryData = {
  industry: GicsIndustry;
  subIndustries: GicsSubIndustryData[];
};

export type GicsIndustryGroupData = {
  industryGroup: GicsIndustryGroup;
  industries: GicsIndustryData[];
};

export type GicsSectorData = {
  sector: GicsSector;
  industryGroups: GicsIndustryGroupData[];
};